after writing this database test. It seems my db logic is not sound.  I believe I may be feeding my network null images, which is resulting in almost random behavior.
If two null images are applied to different labels, my guess is it will throw off the network, which is why its trending towards zero. This seems to be happening with the host version of the database, which means it is also probably happening with the cuda version since the cuda version is copied from the host version. It's weird because when i was requesting images one at a time by asking for user input, I was not getting any errors. Also the test is not failing. The program is terminating early, without throwing errors. I should add debugging flags to the makefile.